+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-08-07
+
+### Added
+
+- Workstones now support dispensers.
+- Hammers can now be displayed on Workstones (shift+right click).
+- Added more action bar feedback for invalid Workstone recipes.
+- Added hammering sounds for hammered blocks.
+
+### Changed
+
+- Workstone plate pressing now only succeeds 75% of the time (configurable).
+    - Doesn't apply to sequenced assembly recipes.
+- Workstone Comparator output scales based on the stored stack count ratio.
+
 ## [1.0.0] - 2026-08-06
 
 - Initial release.